Move query_from_query_expression over to re_datafusion based on reviewer feedback
