+++ crates/store/re_dataframe/Cargo.toml
@@ -31,7 +31,6 @@ re_chunk.workspace = true
 re_format_arrow.workspace = true
 re_log_types.workspace = true
 re_log.workspace = true
-re_protos.workspace = true
 re_query.workspace = true
 re_sorbet.workspace = true
 re_tracing.workspace = true
