+++ crates/store/re_datafusion/src/lib.rs
@@ -12,7 +12,7 @@ mod search_provider;
 mod table_entry_provider;
 mod wasm_compat;
 
-pub use dataframe_query_common::DataframeQueryTableProvider;
+pub use dataframe_query_common::{DataframeQueryTableProvider, query_from_query_expression};
 #[cfg(not(target_arch = "wasm32"))]
 pub(crate) use dataframe_query_provider::PartitionStreamExec;
 #[cfg(target_arch = "wasm32")]
